refactor: separate page and view ApperanceSettings
